fix map won't initialize in the site editor
the ussue was caused by the mapkit js checking wether the map node is an instance of window.Node. Which it isn't since it was generated by a different window in the iframe.

Therefore we now proxy it through a new node generated by the corect window and passing that to mapkit. After that we insert that node into the original node.